% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oexp_read.R
\name{oexp_read}
\alias{oexp_read}
\title{Read OSM data}
\usage{
oexp_read(path, expand_tags = TRUE)
}
\arguments{
\item{path}{Path to .osm file (string)}

\item{expand_tags}{Should the "other_tags" columns be expanded into single-tag columns? (logical; default is TRUE)}
}
\value{
An object of class "osm". There are print and plot methods for this
class.
}
\description{
Read .osm files downloaded from the "Export" page on openstreetmap.org.
}
