% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_separate_tags.R
\name{osm_separate_tags}
\alias{osm_separate_tags}
\title{Separate other_tags column into separate, single-tag columns}
\usage{
osm_separate_tags(x, col_name = "other_tags")
}
\arguments{
\item{x}{sf object containing a column grouping many OSM tags}

\item{col_name}{name of the column containing grouped tags (string; default is "other_tags")}
}
\value{
sf object with single-tag columns
}
\description{
Separate other_tags column into separate, single-tag columns
}
