% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_separate_tags.R
\name{ox_separate_tags}
\alias{ox_separate_tags}
\title{Separate other_tags column into separate, single-tag columns}
\usage{
ox_separate_tags(x, col_name = "other_tags")
}
\arguments{
\item{x}{sf object containing a column grouping many OSM tags}

\item{col_name}{name of the column containing grouped tags (string; default is "other_tags")}
}
\value{
sf object with single-tag columns
}
\description{
Separate other_tags column into separate, single-tag columns
}
\examples{
sb <- ox_read(
  system.file("extdata/south_brisbane.osm",
              package = "osmxml"),
  expand_tags = FALSE)
ox_separate_tags(sb$points)
}
