% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.osm.R
\name{plot.osm}
\alias{plot.osm}
\title{Plot OSM data}
\usage{
\method{plot}{osm}(
  x,
  layers = c("multipolygons", "multilinestrings", "lines", "points", "other_relations"),
  palette = c(1, 2, 3, 4, 6)
)
}
\arguments{
\item{x}{Object of class 'osm'}

\item{layers}{Character vector of names of items to plot}

\item{palette}{Vector of colour values of same length as layers}
}
\value{
A static visualisation
}
\description{
Rudimentary rendering of an OSM export, with one style per layer.
}
