% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_osm.R
\name{read_osm}
\alias{read_osm}
\title{Read OSM data}
\usage{
read_osm(path)
}
\arguments{
\item{path}{Path to .osm file (string)}
}
\value{
An object of class "osm". There are print and plot methods for this
class.
}
\description{
Read .osm files downloaded from the "Export" page on openstreetmap.org.
}
