% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oexp_download.R
\name{oexp_download}
\alias{oexp_download}
\title{Download OSM export}
\usage{
oexp_download(bbox, destfile = "map.osm")
}
\arguments{
\item{bbox}{bounding box of area to download, determined by a numeric vector
of four values: Western longitude, Southern latitude, Eastern longitude,
Northern latitude (in this order)}

\item{destfile}{path to download the file to. Defaults to "map.osm" in the
current working directory.}
}
\value{
The path is returned so it can easily be piped into the
  \code{\link{oexp_read}} function
}
\description{
Download an OSM export by specifying its bounding box, which only mimics the
\url{https://www.openstreetmap.org} "Export" functionality. The
osmexport package does not further make use of the OSM API, given
that it is mainly intended to be used for editing purposes, as explained on
the API Usage Policy website:
\url{https://operations.osmfoundation.org/policies/api/}
}
\section{Licence}{


All data downloaded with this function is © OpenStreetMap contributors,
and the conditions of its reuse are defined by the ODbL licence.
You likely need to include this information on anything derived from it.
Find out more on the OSM website: \url{https://www.openstreetmap.org/copyright}
}

\examples{
\dontrun{
# Download and read area around Te Kura Tatauranga, Waipapa Taumata Rau
TKT <- c(174.76598, -36.85440, 174.77019, -36.85129) |>
   oexp_download() |>
   oexp_read()
}
}
