% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_read.R
\name{print.osm}
\alias{print.osm}
\title{Print method for OSM objects}
\usage{
\method{print}{osm}(obj)
}
\arguments{
\item{obj}{Object of class "osm"}
}
\value{
Description of object: what it is and what it contains.
}
\description{
Print method for OSM objects
}
\examples{
ox_read(
  system.file("extdata/south_brisbane.osm",
              package = "osmxml"),
  expand_tags = FALSE) |>
  print()
}
