% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_osm.R
\name{print.osm}
\alias{print.osm}
\title{Print method for OSM objects}
\usage{
\method{print}{osm}(obj)
}
\arguments{
\item{obj}{Object of class "osm"}
}
\value{
Description of object: what it is and what it contains.
}
\description{
Print method for OSM objects
}
